logged out validation for cart and remove cart cart update cart added

+++ src/Components/CartAside/CartAside.js
@@ -32,8 +32,8 @@ export const CartAside = ({ setCartPop }) => {
         } else {
             let newQty = currQty - 1;
             if (newQty === 0) {
-                if(appData.appData.cartSaved === true){
-                    let cartID = cartInfo[cartProdID].cart_id;
+                let cartID = cartInfo[cartProdID].cart_id;
+                if(appData.appData.cartSaved === true && cartID !== null && cartID != undefined){
                     const payload = {
                         store_id: enviroment.STORE_ID,
                         customer_id: userInfo.customer_id,
@@ -62,16 +62,18 @@ export const CartAside = ({ setCartPop }) => {
         if(appData.appData.cartSaved === true){
             let cartProdID = cartInfo.findIndex((obj) => obj.product_id === id);
             let cartID = cartInfo[cartProdID].cart_id;
-            const payload = {
-                store_id: enviroment.STORE_ID,
-                customer_id: userInfo.customer_id,
-                cart_id: cartID
+            if(appData.appData.cartSaved === true && cartID !== null && cartID != undefined){
+                const payload = {
+                    store_id: enviroment.STORE_ID,
+                    customer_id: userInfo.customer_id,
+                    cart_id: cartID
+                }
+                ApiService.removeCart(payload).then((res) => {
+
+                }).catch((err) => {
+
+                });
             }
-            ApiService.removeCart(payload).then((res) => {
-
-            }).catch((err) => {
-
-            });
         }
         let newCartInfo = cartInfo.filter((obj) => obj.product_id !== id);
         cartInfo = newCartInfo;
@@ -99,6 +101,11 @@ export const CartAside = ({ setCartPop }) => {
         setCartData(appData?.appData?.cartData);
         setCartTotal(appData?.appData?.cartData);
     }, []);
+
+    useEffect(() => {
+        setCartData(appData?.appData?.cartData);
+        setCartTotal(appData?.appData?.cartData);
+    }, [appData?.appData]);
 
     return (
         <React.Fragment>
